auth-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
asyncpg==0.29.0